chore: Update package.json scripts, fix Elysia routes, and remove unused code
